fix(geometry) syncronized geometry with math (previously moved)
