+++ src/core/geometry/geom2/union.js
@@ -1,4 +1,4 @@
-const retessellate = require('../shape3/retessellate')
+const retessellate = require('../geom3/retessellate')
 const canonicalize = require('./canonicalize')
 const toShape3Wall = require('./toGeom3Wall')
 const fromFakeShape3 = require('./fromFakeGeom3')
