+++ src/core/geometry/geom3/measureBounds.js
@@ -1,4 +1,4 @@
-const vec3 = require('../../math/vec3')
+const vec3 = require('../../../math/vec3')
 const measureBoundingBox = require('../poly3/measureBoundingBox')
 
 /**
