+++ src/core/geometry/poly3/fromPoints.js
@@ -1,7 +1,7 @@
 const create = require('./create')
 
-const plane = require('../../math/plane/')
-const vec3 = require('../../math/vec3')
+const plane = require('../../../math/plane/')
+const vec3 = require('../../../math/vec3')
 
 /**
  * Create a poly3 from the given points.
