+++ src/core/geometry/poly3/measureArea.test.js
@@ -1,7 +1,7 @@
 const test = require('ava')
 const { measureArea, create, flip, fromPoints, transform } = require('./index')
 
-const mat4 = require('../../math/mat4')
+const mat4 = require('../../../math/mat4')
 
 const { nearlyEqual } = require('../../../../test/helpers/index')
