+++ src/core/geometry/poly3/measureSignedVolume.js
@@ -1,4 +1,4 @@
-const vec3 = require('../../math/vec3')
+const vec3 = require('../../../math/vec3')
 
 /*
  * Measure the signed volume of the given polygon, which must be convex.
