+++ src/core/geometry/poly3/toString.js
@@ -1,5 +1,5 @@
-const plane = require('../../math/plane/')
-const vec3 = require('../../math/vec3/')
+const plane = require('../../../math/plane/')
+const vec3 = require('../../../math/vec3/')
 
 const toString = (poly3) => {
   let result = `poly3: plane: ${plane.toString(poly3.plane)}, vertices: [`
