+++ src/core/geometry/poly3/translate.js
@@ -1,6 +1,6 @@
 const transform = require('./transform')
 
-const mat4 = require('../../math/mat4')
+const mat4 = require('../../../math/mat4')
 
 const translate = (vector, poly3) => {
   return transform(mat4.fromTranslation(vector), poly3)
